Fix problem on pycbc live due to numpy 2  (#4914)
* Modify the sanity checks of the PSD to be compatible with numpy 2.x
and add sanity check on the PSD of the XML file.
Modify gracedb.py to be up to date with numpy2.x

* New commit to add the import to pass test

* Correcting the import of LigoLWEventSource

* Satisfy codeclimate

---------

+++ pycbc/io/gracedb.py
@@ -222,7 +222,10 @@ class CandidateForGraceDB(object):
             fseries = lal.CreateREAL8FrequencySeries(
                 "psd", psd.epoch, kwargs['low_frequency_cutoff'], psd.delta_f,
                 lal.StrainUnit**2 / lal.HertzUnit, len(psd) - kmin)
-            fseries.data.data = psd.numpy()[kmin:] / pycbc.DYN_RANGE_FAC ** 2.0
+            fseries.data.data = (
+                psd.numpy()[kmin:].astype(numpy.float64)
+                / pycbc.DYN_RANGE_FAC ** 2.0
+            )
             psds_lal[ifo] = fseries
         make_psd_xmldoc(psds_lal, outdoc)
